# Improve performance of DateTime.Kind property.