--- a/src/libraries/System.Private.CoreLib/src/System/DateTime.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/DateTime.cs
@@ -1047,14 +1047,17 @@ namespace System
         internal bool IsAmbiguousDaylightSavingTime() =>
             InternalKind == KindLocalAmbiguousDst;
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public DateTimeKind Kind =>
-            InternalKind switch
+
+        public DateTimeKind Kind
+        {
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            get => InternalKind switch
             {
                 KindUnspecified => DateTimeKind.Unspecified,
                 KindUtc => DateTimeKind.Utc,
                 _ => DateTimeKind.Local,
             };
+        }
 
         // Returns the millisecond part of this DateTime. The returned value
         // is an integer between 0 and 999.