Yesterday, I stumbled upon the fact that `DateTime.Kind` does not get inlined and generally produces unexpectedly large code footprint when used. I thus had to basically "manually inline" it in my library, which is not ideal. 

This PR focuses on bringing some improvements to `DateTime.Kind` with the aim of reducing the machine code size and time it takes to calculate the property. This comes with a drawback: the new algorithm is not as readable as the one used before and also depends on the definition of `DateTimeKind` members. To that end, I placed a large comment explaining what is going on in the getter. 

[Here](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABAJgEYBYAKGIAYACY8gOgDkBXfGKASzFwBuGjWIBmBrgxQOYDAwAi2DDBoBvGgy0MADnwBuymA1jYAJhAB2AGwCeDDtasBzBgH0zy7MOraGm7UhLKQcnS1cAaV5LMwBVYJ0YMF4AM14YMwYAXgY6BDoCwqLiuh8/IJDHFwYomNiMMGzchBQStoKfAEgK+SrwhgAxa2xnXABZPABrJryAYXa2su09XkMVUOqASUsVKEtsa1rMrIA+d08MbAYAMkHh0YncSaWtAK1xGujMo+yz7d39ocvpIAO68BoACzeDA0vj82iO8VwiWSaQyv0+MRYSJRqXSZjQ0L8iIaGKO2IahLh8PcZK+LAAMhAwAdoQBfF7+alMCQ/AZ4DBHAAUAEpobDOp1DFAGJNgTkhUcRUKPF4GCczkhSCKulLsDLeLgmSzrE05TF1ZizIzmQddakGELDcaDmLqJK1PCaJLOubjlabSbdWzvd7OsQAOyOpV+zkh6jxvyiCQwSxcK3qaE4pJ49E5OhUvz1Ro5ciF7Qu005Ug0NlAA), I recreated the "before" and "after" state on Sharplab: the code is 3 times smaller and has one less branch.

I also measured the execution time. The upper bound on the improvement is ~6x on my machine.
<details>
<summary>Benchamarks</summary>

```C#
public class Benchmarks
{
    public const ulong N = 100000;
    
    [Benchmark(Baseline =true)]
    public Kind SwitchKind()
    {
        var result = Kind.Unspecified;
        for (ulong i = 0; i < N; i++)
        {
            result = Unsafe.As<ulong, Date>(ref i).Kind;
        }

        return result;
    }

    [Benchmark]
    public DateTimeKind BitKind()
    {
        var result = DateTimeKind.Unspecified;
        for (ulong i = 0; i < N; i++)
        {
            result = Unsafe.As<ulong, Date>(ref i).BitKind;
        }

        return result;
    }
}

public struct Date
{
    private readonly ulong _data;

    private const ulong KindUnspecified = 0x0000000000000000;
    private const ulong KindUtc = 0x4000000000000000;
    private const ulong FlagsMask = 0xC000000000000000;

    public Date(ulong data) => _data = data;

    public Kind Kind => InternalKind switch
    {
        KindUnspecified => Kind.Unspecified,
        KindUtc => Kind.Utc,
        _ => Kind.Local
    };

    public Kind BitKind
    {
        get
        {                
            var kind = (Kind)(_data >> 62);
            var isAmbiguousLocal = kind > Kind.Local;
            if (isAmbiguousLocal)
            {
                kind = Kind.Local;
            }

            return kind;
        }
    }

    private ulong InternalKind => _data & FlagsMask;
}

public enum Kind
{
    Unspecified = 0,
    Utc = 1,
    Local = 2,        
}
```
</details>

<details>
<summary>Results</summary>

> BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.836 (1909/November2018Update/19H2)
Intel Core i7-4820K CPU 3.70GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.301
  [Host]     : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT


|     Method |      Mean |    Error |   StdDev | Ratio |
|----------- |----------:|---------:|---------:|------:|
| SwitchKind | 389.18 us | 6.189 us | 5.789 us |  1.00 |
|    BitKind |  55.51 us | 0.701 us | 0.621 us |  0.14 |
</details>

If it's needed, I will try to download the latest preview and tests against that (my internet is very flaky (and capped) so I was kind of holding back on that).