@GrabYourPitchforks 2020-06-28
If you look at [the existing call sites](https://source.dot.net/System.Private.CoreLib/R/9e2babef8039992d.html) of `DateTime.Kind`, almost all of them follow the pattern `if (theDateTime.Kind == <SomeKind>) { /* ... */ }`.

This strongly implies that the change that would have the biggest real-world impact is instead as follows, as it would allow the JIT to prune branches which aren't relevant to the caller.

```cs
[MethodImpl(MethodImplOptions.AggressiveInlining)] // if needed
public DateTimeKind Kind
{
    get
    {
        ulong internalKind = InternalKind;
        if (internalKind == KindUnspecified) { return DateTimeKind.Unspecified; }
        else if (internalKind == KindUtc) { return DateTimeKind.Utc; }
        else { return DateTimeKind.Local; }
    }
}
```