@SingleAccretion 2020-06-28
@GrabYourPitchforks Yes, that is what I tried at first. But it didn't produce method small enough for the .NET Core 3.1 JIT to inline (the code looked a lot like the switch case, actually), so I dropped the idea because `DateTime.Kind` is not a "hot" method and as such I didn't want to use `AgressiveInlining`. 

Now, I recreated all this on [sharplab](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABAJgEYBYAKGIAYACY8gOgDkBXfGKASzFwBuGvSasAShwB2GXtxYBhCPgAOvADY8AyjwBu/GEJoiAzA1wYoHMBgYARbBhg0A3jQYeGKvrscwGsNgAJhBS6gCeDBzqoQDmDAD6QY7YwtSeDO6ekFIWUTFS8QDSvFJBAKq5KjBgvABmvDBBDAC8DHQIdF3dPb10aRk5edFxDCVl5Rhgre0IKH0LXWkAkEO2I4UMAGLq2LG4ALJ4ANYzHQqLCwOe3ry+TvmjAJIyPFLY6uPNLQB8ickYbAMABk212+yOuGO1w8WQ8xDMXwYWgA7rwpgALL4ACgAlK0/i8nFB3p9Ss1cGjMXCGG50hlPF9KrhqrUGk0CWNySxmaz6o0gmgaRkmVNOV8eVMhfSGYlxdyADIQMAfGkAXxhmRlTER5IYACF0TjcTS6ctlr4oAxjnq2tivrjsUkUgwfn8kKRcSsLdgrbxcEqVeoZjayq6uWUWIGPt76gxsf7o+oTdRzS4GTRzctQ98I0Eo8qY5nlmrM8XiAB2a3kzWl7U0gDaBxgGAxECCT1U6mxzdb7c7KnUAHkVLJQrgWABBWKxWC4XB3GAvdSlUqxXEAXRpCLzDCedTq4S+y9KTTxptlsO1GQ28VKxNJSLaRLeHy+moycYTrxJb9tbSZKoan5Jp8XTSs8x5IC2QFQQGDrS8GBgdRcH8L971fMkwxaADyUmMAwKYKsJXwuCEMvZDUNpIjIKTMjEPVYVL23XUw33Q9jQvZjr08W8GAw38sNzF9BPfJiPHQn9H3/PNeWA9kgkIiCSOgkCgjI8SkJQtC6njATpOw3CJimJTiO5Uj4M0yj/HAszIzo+CGO1cirwyHdgAgCBgyeXBUXRMAsXJJNsQcB4ARgfFfgYcKWD8zFjVaIz8zomkWIYDyvL3XBDQwL5gtC/xwsiv4YpyhKcNowt1E1NKMu83ATykJp2KPIKqpCvxor8YquqcFgWuPMJTyCPFEsqoMau1dzPPqga2qDDqwu6zkYrmspRoqiUUuoFytR3GApC4PNXBpOSYI5No6GlDJ8Jmchrs8JMZlIGg1SAA), and the results (unfortunately) do still favor my version, even with the aggressive inlining. 

I will investigate further to see if it is actually possible to coerce JIT into doing something better here.