@SingleAccretion 2020-06-28
After some playing around, I was able to construct a case where this version of `DateTime.Kind` with a simple switch is better than mine:
```C#
[MethodImpl(MethodImplOptions.AggressiveInlining)]
public Kind OtherSwitchKind()
{
    var kind = (Kind)(_data >> 62);
    switch (kind)
    {
        case Kind.Unspecified: return Kind.Unspecified;
        case Kind.Utc: return Kind.Utc;            
        default: return Kind.Local;
    }
}
```
With the "benchmark" method being:
```C#
public static int KindTest(Date date)
{
    if (date.GetKind() == Kind.Utc)
    {
        return 7;
    }
    else
    {
        return 90;
    }
}
```
Different variants of `GetKind()` produce ever so slightly different ASM:
```ASM
; Original switch expression with AgressiveInlining
L0000: mov rax, 0xc000000000000000
L000a: and rax, rcx
L000d: test rax, rax
L0010: je short L0027
L0012: mov rdx, 0x4000000000000000
L001c: cmp rax, rdx
L001f: jne short L0027
L0021: mov eax, 7
L0026: ret
L0027: mov eax, 0x5a
L002c: ret 

; This PR's suggested implementation:
L0000: shr rcx, 0x3e
L0004: mov eax, ecx
L0006: cmp eax, 2
L0009: jle short L0010
L000b: mov eax, 2
L0010: cmp eax, 1
L0013: jne short L001b
L0015: mov eax, 7
L001a: ret
L001b: mov eax, 0x5a
L0020: ret

; The switch statement
L0000: shr rcx, 0x3e
L0004: mov eax, ecx
L0006: test eax, eax
L0008: je short L0015
L000a: cmp eax, 1
L000d: jne short L0015
L000f: mov eax, 7
L0014: ret
L0015: mov eax, 0x5a
L001a: ret
```
As can be seen, the differences are literally one 1-2 instructions. This suggests that even just putting `AgressiveInlining` on the property is good enough. I will benchmark the methods and see if there is actually any difference between them.