@GrabYourPitchforks 2020-06-29
Yes, your suggested benchmark method `KindTest` should be a sufficient test. The JIT doesn't always handle _bool_-returning methods all that well, so any test which contains an explicit equality check rather than bouncing through a helper method will be more representative of real-world throughput + codegen.

Also of note is that instruction count alone isn't a good metric of whether a function is fast inside a hot path. On x86, depending on the CPU in use, _shr_ may be 4x as expensive as equivalent _mov_ / _and_ / _test_ instructions. Honestly that's splitting hairs, but in a _reeeeeeally_ tight path it's an interesting example of code size / locality vs. raw throughput as a tradeoff. :)