@GrabYourPitchforks 2020-06-29
And if you really wanted to get crazy:
(This is all completely untested, so I don't know if I screwed up the bit twiddling.)

```cs
    [MethodImpl(MethodImplOptions.AggressiveInlining)]
    public Kind IffyKindInlined3()
    {   
        long data = (long)_data;
        if ((data & unchecked((long)FlagsMask)) == 0) { return Kind.Unspecified; }
        
        // At this point, we know the top 2 bits of 'data' are 0x01, 0x10, or 0x11.
        // The "UTC" pattern sets the top two bits to 0x01. This allows us to get
        // away with a simple "is positive?" check when checking for UTC.
        
        if (data > 0) { return Kind.Utc; }
        else { return Kind.Local; }        
    }
    
    [MethodImpl(MethodImplOptions.AggressiveInlining)]
    public Kind IffyKindInlined4()
    {   
        long data = (long)_data;
        if (data > 0)
        {
            // Top 2 bits of data are 0x00 or 0x01
            data += data;
            if (data < 0)
            {
                return Kind.Utc; // Top 2 bits were 0x01
            } else {
                return Kind.Unspecified; // Top 2 bits were 0x00
            }
        }
        return Kind.Local; // Top 2 bits are 0x10 or 0x11
    }
```

```asm
Date.KindTest3(Date)
    L0000: mov rax, 0xc000000000000000
    L000a: test rax, rcx
    L000d: je short L001a
    L000f: test rcx, rcx
    L0012: jle short L001a
    L0014: mov eax, 7
    L0019: ret
    L001a: mov eax, 0x5a
    L001f: ret

Date.KindTest4(Date)
    L0000: test rcx, rcx
    L0003: jle short L0013
    L0005: add rcx, rcx
    L0008: test rcx, rcx   ; <-- future JIT should elide, see https://github.com/dotnet/runtime/issues/32389
    L000b: jge short L0013
    L000d: mov eax, 7
    L0012: ret
    L0013: mov eax, 0x5a
    L0018: ret
```